Upgrade next-auth and add prisma + migrations
