+++ package.json
@@ -4,12 +4,16 @@
   "private": true,
   "scripts": {
     "dev:mem": "yarn migrate && yarn initialize && NODE_OPTIONS='--inspect --max_old_space_size=4096' next dev",
-    "dev": "yarn migrate && yarn initialize && NODE_OPTIONS='--inspect' next dev",
+    "dev": "yarn migrate && yarn prisma:dev && yarn initialize && NODE_OPTIONS='--inspect' next dev",
     "build": "ISOLATED_INSTANCE=true next build",
-    "start": "yarn migrate && yarn initialize && next start -p ${PORT:-3000}",
+    "start": "yarn migrate && yarn prisma:migrate && yarn initialize && next start -p ${PORT:-3000}",
     "migrate": "cd ./migrations && yarn migrate && cd ..",
+    "migrate:toPrisma": "yarn migrate && prisma migrate resolve --applied 20220506182127_initial",
+    "prisma:migrate": "prisma migrate deploy",
+    "prisma:dev": "prisma migrate dev",
+    "prisma:studio": "prisma studio",
     "initialize": "cd ./flag_initialization && yarn initialize && cd ..",
-    "start:test": "ISOLATED_INSTANCE=true next start",
+    "start:test": "ISOLATED_INSTANCE=true DATABASE_URL=postgres://postgres:password@localhost:5432/formsDB next start",
     "lint": "eslint . ",
     "lint:report": "eslint --output-file eslint_report.json --format json . --ext .js,.jsx,.ts,.tsx",
     "lint:fix": "eslint --fix .",
@@ -29,6 +33,9 @@
     "@aws-sdk/client-s3": "^3.58.0",
     "@aws-sdk/client-sqs": "^3.58.0",
     "@aws-sdk/lib-dynamodb": "^3.58.0",
+    "@next-auth/prisma-adapter": "^1.0.3",
+    "@prisma/client": "^3.12.0",
+    "aws-sdk-client-mock": "^0.6.2",
     "axios": "^0.26.1",
     "classnames": "^2.3.1",
     "cors": "^2.8.5",
@@ -43,11 +50,12 @@
     "markdown-to-jsx": "^7.1.7",
     "mmmagic": "^0.5.3",
     "next": "^12.1.4",
-    "next-auth": "3.29.0",
+    "next-auth": "^4.3.3",
     "next-i18next": "^11.0.0",
     "notifications-node-client": "^5.1.1",
     "pg": "^8.7.3",
     "pino": "^7.9.2",
+    "prisma": "^3.14.0",
     "react": "^18.0.0",
     "react-app-polyfill": "^3.0.0",
     "react-dom": "^18.0.0",
@@ -109,7 +117,9 @@
     "jest": "^27.5.1",
     "jest-css-modules": "^2.1.0",
     "jest-fetch-mock": "^3.0.3",
+    "jest-mock-extended": "^2.0.5",
     "node-mocks-http": "^1.11.0",
+    "pg-server": "^0.1.5",
     "pino-pretty": "^7.6.0",
     "postcss": "^8.4.12",
     "postcss-import": "^14.1.0",
@@ -119,6 +129,7 @@
     "sass-loader": "^12.6.0",
     "tailwindcss": "^3.0.23",
     "ts-jest": "^27.1.4",
+    "ts-node": "^10.7.0",
     "typescript": "^4.6.3",
     "webpack": "^5.71.0"
   },
