+++ prisma/schema.prisma
@@ -1,0 +1,95 @@
+generator client {
+  provider = "prisma-client-js"
+}
+
+datasource db {
+  provider = "postgresql"
+  url      = env("DATABASE_URL")
+}
+
+model Account {
+  id                 String  @id @default(cuid())
+  userId             String
+  type               String
+  provider           String
+  providerAccountId  String
+  refresh_token      String?  @db.Text
+  access_token       String?  @db.Text
+  expires_at         Int?
+  token_type         String?
+  scope              String?
+  id_token           String?  @db.Text
+  session_state      String?
+
+  user             User   @relation(fields: [userId], references: [id], onDelete: Cascade, onUpdate: NoAction)
+
+  @@unique([provider, providerAccountId])
+}
+
+model AdminLog {
+  id          Int      @id @default(autoincrement())
+  userId     String
+  timestamp   DateTime @default(now()) @db.Timestamptz(6)
+  action      String   
+  event       String   
+  description String?
+  user       User    @relation(fields: [userId], references: [id], onDelete: NoAction, onUpdate: NoAction)
+}
+
+model FormUser {
+  id              String       @id @default(cuid())
+  templateId     String
+  email           String    @db.VarChar(80)
+  temporaryToken String?
+  active          Boolean   @default(true)
+  created_at      DateTime  @default(now()) @db.Timestamptz(6)
+  updated_at      DateTime  @default(now()) @db.Timestamptz(6)
+  template       Template @relation(fields: [templateId], references: [id], onDelete: NoAction, onUpdate: NoAction)
+
+  @@unique([templateId, email])
+}
+
+model Organization {
+  id     String  @id @default(cuid())
+  nameEn String? 
+  nameFr String? 
+  users  User[]
+}
+
+model Session {
+  id           String   @id @default(cuid())
+  userId       String
+  expires      DateTime
+  sessionToken String @unique(map: "session_token")
+  user        User  @relation(fields: [userId], references: [id], onDelete: Cascade, onUpdate: NoAction)
+}
+
+model Template {
+  id           String         @id @default(cuid())
+  jsonConfig  Json?
+  organization String?      @db.Uuid
+  bearerToken String?
+  formUsers   FormUser[]
+}
+
+model User {
+  id            String            @id @default(cuid())
+  name          String?
+  email         String?        @unique
+  emailVerified DateTime?
+  image         String?
+  admin         Boolean?       @default(false)
+  organizationId  String?        
+  organization Organization? @relation(fields: [organizationId], references: [id], onDelete: NoAction, onUpdate: NoAction)
+  accounts      Account[]
+  admin_logs    AdminLog[]
+  sessions      Session[]
+}
+
+model VerificationToken {
+  identifier String
+  token      String   @unique
+  expires    DateTime
+
+  @@unique([identifier, token])
+}
